Decompose canvas overlay components
